requirements.txt
pdfplumber>=0.11.0
pytesseract>=0.3.10
pdf2image>=1.17.0
Pillow>=10.3.0
rapidfuzz>=3.9.6
python-dateutil>=2.9.0
pytest>=8.0.0
